Update pet-shop.md
Not sure if it's the right place, but I added App.initContract() to the initWeb3 function because initContract is never called. Without initContract being called, the Adopt buttons don't work and MetaMask doesn't pop up for approval of the adoption transaction. There may be a better way to go about calling initContract, but I figured I'd submit this to highlight the issue.