Remove the word "now" from TS test file docs
Users reading this document have no concept of when the document was written, and so have no point of reference to understand the "now". Additionally, readers unaware of Truffle's past will be reminded of a time Truffle was less awesome, even if they never lived it (e.g., saying something is "now supported" immediately references a time where it wasn't). Instead, we should just remove the "now" and flex our Typescript muscles like it ain't no thang. Feel me? 💪