removed unnecessary properties in action object
Drizzle and web3 were showing as required properties in the action object to be dispatched to add contracts dynamically, these are no longer required. {type: 'ADD_CONTRACT', contractConfig, events}, is all that is needed, and it is more symmetrical with the other methods shown below in line 84, _using the Drizzle context object_ directly.