Update to ES6 and MetaMask new API
Update MetaMask deprecated API -- call to `enable()` is replaced with `request()`.
It is safe to say ES6, in particular `async - await` is a new standard all Javascript users know, and it is way more readable than callbacks and promise chaining.

NOTE: The corresponding functions in petshop box template should be marked `async` too.