Deprecated the Getting Started with Drizzle and React guide
Unfortunately, Drizzle is currently in need of, possibly substantial, work to update
its dependencies. There are presently too many errors that prevent Drizzle from being
useable. As a result, a user attempting to follow this guide will be unable to complete
those portions that require Drizzle, which is the aim of the guide.

The Getting Started with Drizzle and React guide is to be deprecated until such time
as the issues with Drizzle have been resolved. At that time the guide could be
resurrected or a fresh guide prepared.